SurfaceFlinger: only ExplicitDefault can use appRequestRange
This change is limiting layers that are ExplicitExactOrMultiple from
using a refresh rate outside of the primary range. When these layers
are visible and there is an interaction with the device, we usually
change the refresh rate due to other layers that are animating. Letting
layers that are ExplicitExactOrMultiple to pick a refresh rate from the
extended appRequestRange results in refresh rate changes which are not
desired.

Bug: 159940172
Test: YouTube when the device the primary range is restricted
Change-Id: I6aa60c359d690a92342963cb14bdeece4e6d5c5f